requirements.txt
python-binance>=1.0.19
aiohttp>=3.9.5